src/app/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.30.1
python-multipart==0.0.27

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.1
psycopg2-binary==2.9.10

# Redis / Caching
redis==5.0.8

# Settings / Config
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.2.2

# Security
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-jose[cryptography]==3.4.0
cryptography==46.0.7

# HTTP Client
httpx==0.27.0

# Observability
prometheus-client==0.20.0
structlog==24.2.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-exporter-otlp==1.24.0

# Utilities
orjson==3.11.6
tenacity==8.3.0
dnspython==2.8.0
email-validator==2.3.0
